Fix: RootPage 컴포넌트화 및 UI 수정
