+++ fitple/components/Card/Card.tsx
@@ -12,26 +12,21 @@ type CardProps = {
     isLoading?: boolean;
 };
 
-const Card: FC<CardProps> = ({ header, body, footer, size = 'md', isLoading, ...props }) => {
-    const [visible, setVisible] = useState(false);
-
-    useEffect(() => {
-        const timeout = setTimeout(() => {
-            setVisible(true);
-            console.log('셋타임아웃: ', isLoading);
-        }, 10); // 마운트 후 살짝 딜레이 후 트리거
-        return () => clearTimeout(timeout);
-    }, []);
-
+const Card: FC<CardProps> = ({ header, body, footer, size = 'md', isLoading = true, ...props }) => {
+    // useEffect(() => {
+    //     const timeout = setTimeout(() => {
+    //         setVisible(true);
+    //         console.log('셋타임아웃: ', isLoading);
+    //     }, 10); // 마운트 후 살짝 딜레이 후 트리거
+    //     return () => clearTimeout(timeout);
+    // }, []);
+    console.log('값 넘어오나 확인: ', isLoading);
     return (
         <>
-            {isLoading ? (
-                <div>
-                    <SkeletonCard />
-                    ㅎㅇ
-                </div>
+            {isLoading ? ( //true였다가 false로 바뀌면서
+                <SkeletonCard />
             ) : (
-                <div className={`${styles.card} ${styles[size]} ${visible ? styles.show : ''}`} {...props}>
+                <div className={`${styles.card} ${styles[size]} ${isLoading ? '' : styles.show}`} {...props}>
                     <div className={styles.cardHeader}>{header}</div>
                     <div className={styles.cardBody}>{body}</div>
                     {footer && <div className={styles.cardFooter}>{footer}</div>}
